Refactor all the awesomeness
